+++ server/src/test/java/com/ahmadda/domain/notification/ReminderTest.java
@@ -1,4 +1,4 @@
-package com.ahmadda.domain;
+package com.ahmadda.domain.notification;
 
 import com.ahmadda.annotation.IntegrationTest;
 import com.ahmadda.domain.event.Event;
@@ -6,15 +6,11 @@ import com.ahmadda.domain.event.EventOperationPeriod;
 import com.ahmadda.domain.event.EventRepository;
 import com.ahmadda.domain.member.Member;
 import com.ahmadda.domain.member.MemberRepository;
-import com.ahmadda.domain.notification.EmailNotifier;
-import com.ahmadda.domain.notification.PushNotifier;
-import com.ahmadda.domain.notification.Reminder;
-import com.ahmadda.domain.notification.ReminderRecipient;
 import com.ahmadda.domain.organization.Organization;
 import com.ahmadda.domain.organization.OrganizationMember;
 import com.ahmadda.domain.organization.OrganizationMemberRepository;
-import com.ahmadda.domain.organization.OrganizationRepository;
 import com.ahmadda.domain.organization.OrganizationMemberRole;
+import com.ahmadda.domain.organization.OrganizationRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
