+++ server/src/test/java/com/ahmadda/domain/organization/OrganizationMemberTest.java
@@ -1,12 +1,9 @@
-package com.ahmadda.domain;
+package com.ahmadda.domain.organization;
 
 import com.ahmadda.domain.event.Event;
 import com.ahmadda.domain.event.EventOperationPeriod;
 import com.ahmadda.domain.event.Guest;
 import com.ahmadda.domain.member.Member;
-import com.ahmadda.domain.organization.Organization;
-import com.ahmadda.domain.organization.OrganizationMember;
-import com.ahmadda.domain.organization.OrganizationMemberRole;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
