+++ server/src/test/java/com/ahmadda/domain/organization/OrganizationTest.java
@@ -1,14 +1,10 @@
-package com.ahmadda.domain;
+package com.ahmadda.domain.organization;
 
 import com.ahmadda.domain.event.Event;
 import com.ahmadda.domain.event.EventOperationPeriod;
 import com.ahmadda.domain.exception.BusinessRuleViolatedException;
 import com.ahmadda.domain.exception.UnauthorizedOperationException;
 import com.ahmadda.domain.member.Member;
-import com.ahmadda.domain.organization.InviteCode;
-import com.ahmadda.domain.organization.Organization;
-import com.ahmadda.domain.organization.OrganizationMember;
-import com.ahmadda.domain.organization.OrganizationMemberRole;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
