+++ server/src/test/java/com/ahmadda/learning/infra/notification/FcmPushNotifierTest.java
@@ -1,12 +1,12 @@
-package com.ahmadda.learning.notification;
+package com.ahmadda.learning.infra.notification;
 
 import com.ahmadda.annotation.IntegrationTest;
 import com.ahmadda.domain.member.Member;
 import com.ahmadda.domain.member.MemberRepository;
-import com.ahmadda.domain.organization.Organization;
-import com.ahmadda.domain.organization.OrganizationMember;
 import com.ahmadda.domain.notification.PushNotificationPayload;
 import com.ahmadda.domain.notification.PushNotifier;
+import com.ahmadda.domain.organization.Organization;
+import com.ahmadda.domain.organization.OrganizationMember;
 import com.ahmadda.domain.organization.OrganizationMemberRole;
 import com.ahmadda.infra.notification.push.FcmRegistrationToken;
 import com.ahmadda.infra.notification.push.FcmRegistrationTokenRepository;
